improve run.sh a bit for local development
not fetching removing and fetching things all the time

+++ run.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-if [[ -d tmp ]] ; then
+keep_temp=
+if [[ ${1} == "--keep-temp" ]] ; then
+	keep_temp="yes"
+fi
+
+if [[ -d tmp && -z ${keep_temp} ]] ; then
 	rm -r tmp
 fi
 
@@ -10,8 +15,14 @@ echo "date,total_number_of_tests,positivity_rate_percent,isolated,quarantined,so
 mkdir -p tmp
 pushd tmp > /dev/null
 
-wget https://www.bag.admin.ch/dam/bag/de/dokumente/mt/k-und-i/aktuelle-ausbrueche-pandemien/2019-nCoV/bisherige-lageberichte-zip.zip.download.zip/Epidemiologische_Lage_Schweiz.zip
-unzip Epidemiologische_Lage_Schweiz.zip
+if [[ ! -f Epidemiologische_Lage_Schweiz.zip ]] ; then
+	wget https://www.bag.admin.ch/dam/bag/de/dokumente/mt/k-und-i/aktuelle-ausbrueche-pandemien/2019-nCoV/bisherige-lageberichte-zip.zip.download.zip/Epidemiologische_Lage_Schweiz.zip
+fi
+
+if [[ ! -d Deutsch ]] ; then
+	unzip Epidemiologische_Lage_Schweiz.zip
+fi
+
 pushd Deutsch > /dev/null
 
 for i in *.pdf ; do
@@ -20,4 +31,7 @@ done
 
 popd > /dev/null
 popd > /dev/null
-rm -r tmp
+
+if [[ -z ${keep_temp} ]] ; then
+	rm -r tmp
+fi
